Split modules into internal and public folders
