+++ ReactViewResources/Loader/Internal/ViewMetadata.ts
