+++ ReactViewResources/Loader/Internal/ViewPortalsCollection.tsx
